requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
absl-py=0.8.1=py36_0
alabaster=0.7.12=py36_0
argh=0.26.2=py36_0
asn1crypto=1.2.0=py36_0
astor=0.8.0=py36_0
astroid=2.3.3=py36_0
atomicwrites=1.3.0=py36_1
attrs=19.3.0=py_0
autopep8=1.4.4=py_0
babel=2.8.0=py_0
backcall=0.1.0=py36_0
bcrypt=3.1.7=py36he774522_0
blas=1.0=mkl
bleach=3.1.0=py36_0
ca-certificates=2019.11.27=0
certifi=2019.11.28=py36_0
cffi=1.13.2=py36h7a1dbc1_0
chardet=3.0.4=py36_1003
cloudpickle=1.2.2=py_0
colorama=0.4.3=py_0
cryptography=2.8=py36h7a1dbc1_0
cycler=0.10.0=py36h009560c_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
diff-match-patch=20181111=py_0
docutils=0.15.2=py36_0
entrypoints=0.3=py36_0
flake8=3.7.9=py36_0
freetype=2.9.1=ha9979f8_1
future=0.18.2=py36_0
gast=0.2.2=py36_0
google-pasta=0.1.8=py_0
grpcio=1.16.1=py36h351948d_1
h5py=2.8.0=py36hf7173ca_2
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=py36_0
imagesize=1.2.0=py_0
importlib_metadata=1.3.0=py36_0
intel-openmp=2019.4=245
intervaltree=3.0.2=py_0
ipykernel=5.1.3=py36h39e3cac_0
ipython=7.11.1=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.5.1=py_0
isort=4.3.21=py36_0
jedi=0.14.1=py36_0
jinja2=2.10.3=py_0
joblib=0.14.1=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py36_0
jupyter=1.0.0=py36_7
jupyter_client=5.3.4=py36_0
jupyter_console=6.0.0=py36_0
jupyter_core=4.6.1=py36_0
keras=2.2.4=0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py36_0
keras-preprocessing=1.1.0=py_1
keyring=20.0.0=py36_0
kiwisolver=1.1.0=py36ha925a31_0
lazy-object-proxy=1.4.3=py36he774522_0
libmklml=2019.0.5=0
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.11.2=h7bd577a_0
libsodium=1.0.16=h9d3ae62_0
libspatialindex=1.9.3=h33f27b4_0
libtiff=4.1.0=h56a325e_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.1.1=py36_0
markupsafe=1.1.1=py36he774522_0
matplotlib=3.1.1=py36hc8f65d3_0
mccabe=0.6.1=py36_1
mistune=0.8.4=py36he774522_0
mkl=2019.4=245
mkl-service=2.3.0=py36hb782905_0
mkl_fft=1.0.15=py36h14836fe_0
mkl_random=1.1.0=py36h675688f_0
more-itertools=8.0.2=py_0
msys2-conda-epoch=20160418=1
nbconvert=5.6.1=py36_0
nbformat=4.4.0=py36_0
notebook=6.0.2=py36_0
numpy=1.17.4=py36h4320e6b_0
numpy-base=1.17.4=py36hc3f5095_0
numpydoc=0.9.2=py_0
opencv=3.4.2=py36h40b0b35_0
openssl=1.1.1d=he774522_3
opt_einsum=3.1.0=py_0
packaging=20.0=py_0
pandas=0.25.3=py36ha925a31_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
paramiko=2.6.0=py36_0
parso=0.5.2=py_0
pathtools=0.1.2=py_1
pexpect=4.7.0=py36_0
pickleshare=0.7.5=py36_0
pip=19.3.1=py36_0
pluggy=0.13.1=py36_0
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.10=py_0
protobuf=3.11.2=py36h33f27b4_0
psutil=5.6.7=py36he774522_0
py-opencv=3.4.2=py36hc319ecb_0
pybluez=0.23=pypi_0
pycodestyle=2.5.0=py36_0
pycparser=2.19=py36_0
pydocstyle=4.0.1=py_0
pyflakes=2.1.1=py36_0
pygments=2.5.2=py_0
pylint=2.4.4=py36_0
pynacl=1.3.0=py36h62dcd97_0
pyopenssl=19.1.0=py36_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py36h6538335_2
pyrsistent=0.15.6=py36he774522_0
pyserial=3.4=py36_0
pysocks=1.7.1=py36_0
python=3.6.10=h9f7ef89_0
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.3.2=py_0
python-language-server=0.31.2=py36_0
pytz=2019.3=py_0
pywin32=227=py36he774522_0
pywin32-ctypes=0.2.0=py36_0
pywinpty=0.5.7=py36_0
pyyaml=5.2=py36he774522_0
pyzmq=18.1.0=py36ha925a31_0
qdarkstyle=2.7=py_0
qt=5.9.7=vc14h73c81de_0
qtawesome=0.6.0=py_0
qtconsole=4.6.0=py_1
qtpy=1.9.0=py_0
requests=2.22.0=py36_1
rope=0.14.0=py_0
rtree=0.8.3=py36_0
scikit-learn=0.22.1=py36h6288b17_0
scipy=1.3.2=py36h29ff71c_0
send2trash=1.5.0=py36_0
setuptools=44.0.0=py36_0
sip=4.19.8=py36h6538335_0
six=1.13.0=py36_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.1.0=py36_0
sphinx=2.3.1=py_0
sphinxcontrib-applehelp=1.0.1=py_0
sphinxcontrib-devhelp=1.0.1=py_0
sphinxcontrib-htmlhelp=1.0.2=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.2=py_0
sphinxcontrib-serializinghtml=1.1.3=py_0
spyder=4.0.1=py36_0
spyder-kernels=1.8.1=py36_0
sqlite=3.30.1=he774522_0
tensorboard=2.0.0=pyhb38c66f_1
tensorflow=1.15.0=mkl_py36h997801b_0
tensorflow-base=1.15.0=mkl_py36h190a33d_0
tensorflow-estimator=1.15.1=pyh2649769_0
termcolor=1.1.0=py36_1
terminado=0.8.3=py36_0
testpath=0.4.4=py_0
tornado=6.0.3=py36he774522_0
traitlets=4.3.3=py36_0
typed-ast=1.4.0=py36he774522_0
ujson=1.35=py36hfa6e2cd_0
urllib3=1.25.7=py36_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
watchdog=0.9.0=py36_1
wcwidth=0.1.7=py36_0
webencodings=0.5.1=py36_1
werkzeug=0.16.0=py_0
wheel=0.33.6=py36_0
widgetsnbextension=3.5.1=py36_0
win_inet_pton=1.1.0=py36_0
wincertstore=0.2=py36h7fe50ca_0
winpty=0.4.3=4
wrapt=1.11.2=py36he774522_0
xz=5.2.4=h2fa13f4_4
yaml=0.1.7=hc54c509_2
yapf=0.28.0=py_0
zeromq=4.3.1=h33f27b4_3
zipp=0.6.0=py_0
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
